SharedWorker: Factor out WorkerReportingProxy impl from WebSharedWorkerImpl
WebSharedWorkerImpl is a heavily threaded class. Some methods are called on the
main thread, while others are called on a worker thread. This often confuses my
head...

This CL factors out a part of WebSharedWorkerImpl that implement
WorkerReportingProxy and are called on a worker thread into
WebSharedWorkerReportingProxyImpl. Even with this cleanup, some methods called
on a worker thread still remain in WebSharedWorkerImpl, but this would somewhat
mitigate the situation and make threading boundary clearer.

BUG=673711

Review-Url: https://codereview.chromium.org/2721543002
Cr-Original-Commit-Position: refs/heads/master@{#453499}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0decaefae68c308e45aad3b3fbda9c5b9a6793a8