Remove third_party/libjingle and update GN targets.
In order to keep actual changes to a minimum, this CL
just moves the targets:
//third_party/libjingle:libjingle -> //third_party/webrtc_overrides
//third_party/libjingle/webrtc:libjingle_webrtc -> //third_party/webrtc_overrides:init_webrtc
and updates the paths to the webrtc_overrides sources.

References to the old targets are updated to the new ones, with no
additional changes.

The //third_party/libjingle:libstunprober target is no longer referenced
in Chrome (it's been replaced by //third_party/webrtc/p2p:libstunprober)
so it was removed.

BUG=677948

Review-Url: https://codereview.chromium.org/2413103004
Cr-Original-Commit-Position: refs/heads/master@{#453530}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3cbd6d64e02369e691f65f791ab99258305243d