Move task.describe descriptiosn to audit.define
Move the description of the task into the first argument
audit.define.  This is the first step at removing support for
task.describe.

The conversion was done using the gawk script:

/^ *audit.define\(/ {
  audit = $0
  next
}
/^ *task.describe\(.*\);/ {
  desc_match = match($0, /^( *)task.describe\((.*)\)/, desc_pat);
  label_match = match(audit, /^( *)audit.define\((['"].*['"]),(.*)/, label_pat)
  printf "%saudit.define({\n", label_pat[1]
  printf "%s  label: %s,\n", label_pat[1], label_pat[2]
  printf "%s  description: %s\n", label_pat[1], desc_pat[2]
  printf "%s},%s", label_pat[1], label_pat[3]
  next
}
{ print }

This script tries to indent things reasonably nicely, but doesn't
handle the case where the description is very long or where the indentation is different.

Also, not all files could be converted automatically because the
script is not smart enough.  These will be handled separately.

BUG=694659
TEST=none

Review-Url: https://codereview.chromium.org/2708953003
Cr-Original-Commit-Position: refs/heads/master@{#453770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ae3ac58c9bc4d6c20a3ec2dab0b08dd62acb1dc