Use the Oilpan heap canary only on 64-bit.
On 64-bit machines, the canary is free, space-wise. But on 32-bit, it doubles
the size of |HeapObjectHeader|, which has incurred a large run-time memory size
increase.

Turn off the canary for now, on 32-bit. Ultimately we may solve the original bug
(633030) by moving the metadata out of line instead of guarding it with a
canary.

BUG=692931,692624,633030

Review-Url: https://codereview.chromium.org/2723743003
Cr-Original-Commit-Position: refs/heads/master@{#453802}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3d585cce44455344c067ab7a5ebb359fae854f7