Implement 'normal', 'strict', and 'loose' of the 'line-break' property
This patch implements the 3 values of the 'line-break' property[1] using
the CLDR text break variants in locale IDs[2]. This is the same way as
WebKit supports these 3 values.

This patch fixes 82 test failures in csswg-test. 43 tests still fail,
they are to be investigated.

This patch relies on ICU to have necessary data. ICU is working on its
efficiency further[3].

[1] https://drafts.csswg.org/css-text-3/#line-break-property
[2] http://cldr.unicode.org/development/development-process/design-proposals/specifying-text-break-variants-in-locale-ids
[3] http://bugs.icu-project.org/trac/ticket/11530

BUG=306730

Review-Url: https://codereview.chromium.org/2715153003
Cr-Original-Commit-Position: refs/heads/master@{#453809}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 73f7033f99cd065bb7fb7531c8dc3cb2b115fd5e