Renamed deviceScaleFactor() to deviceScaleFactorDeprecated() in Page
Renamed the deviceScaleFactor() method to deviceScaleFactorDeprecated()
in the Page class, to be consistent with the FrameHost method that calls
this one. Also moved the comment for the method to the Page class.

This is pre-work to remove FrameHost::deviceScaleFactorDeprecated(), and
is part of merging FrameHost back into Page.

BUG=691794
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2724543002
Cr-Original-Commit-Position: refs/heads/master@{#453871}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0c23baccef3a9647076ad0dde83a1a847c702b03