scheduler: Ensure consistent delayed task ordering between task queues
When wakeups are scheduled for distinct task queues, they should be
ordered according to the tasks which caused the wakeups. Previously this
was only being done by looking at the the scheduled run times fo the
tasks, but we should also consider the sequence numbers for tasks with
identical run times.

BUG=696001

Review-Url: https://codereview.chromium.org/2718293003
Cr-Original-Commit-Position: refs/heads/master@{#453927}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8579f022a226b0182ba7c5a0a2b83815775a1faa