Implements CSSPropertyAPI for the offset-anchor property.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyOffsetAnchor from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
offset-anchor property is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIOffserAnchor.cpp to the BUILD.gn file.
- Adds api_class to CSSProperties.in, which indicates that
  CSSPropertyAPIOffsetAnchor.h is generated.
- Adds api_methods to CSSProperties.in, which defines which methods
  are implemented in the OffserAnchor API
- Moves the parsing logic for offset-anchor from CSSPropertyParser.cpp
  to CSSPropertyAPIOffsetAnchor.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2717373004
Cr-Original-Commit-Position: refs/heads/master@{#454120}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4f9ef034e5650ae8a966baadc47f3a29f3f8edf