Add UMA histogram for media control download button.
Whenever we display media elements, we sometimes display a download
button. This CL adds a histogram for when the media controls display a
download button and when the button is clicked.

It ensures that for each instance of the download button, shown and
clicked can each only be recorded once.

BUG=None

Review-Url: https://codereview.chromium.org/2723293002
Cr-Original-Commit-Position: refs/heads/master@{#454256}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b41472a230a8cf773aeb018177f83032498da650