Remove non-generated member 'm_unique' from ComputedStyle.
In a previous patch (crrev.com/2667543002), the nonproperty 'unique'
became generated in ComputedStyleBase. However, the original handwritten
'unique' was not removed from ComputedStyle, so we currently have two
declarations of 'unique', one of which is not used.

This patch removes the handwritten version which is unused.

BUG=628043

Review-Url: https://codereview.chromium.org/2725853003
Cr-Original-Commit-Position: refs/heads/master@{#454269}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 600391d1b460f14000ea7035144f1d5d3460a9a5