[LayoutNG] Snap the baseline position to CSS pixel
This patch rounds down the top half leading so that the top of text box
aligns to CSS pixel. This matches the positioning of text to the legacy
layout when the half leading is fractional and ascent/descent are not,
while keeping enough line height for fractional height.

How to handle fractional ascent and descent needs more experiments and
thoughts.

BUG=636993

Review-Url: https://codereview.chromium.org/2728643005
Cr-Original-Commit-Position: refs/heads/master@{#454323}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5cb819e3a672f5902622a23b456947713f405fa9