PlzNavigate: Rewrite the url and frame type validation part of the  external/wpt/service-workers/service-worker/navigate-window.https.html layout test.
This test expects the ServiceWorkerProviderHosts to be in a specific order which breaks with PlzNavigate due to precreated hosts. It seemed to be a better
to rewrite the validation code to do an exhaustive match of the URLs and the frame types in the result list.

BUG=697307

Review-Url: https://codereview.chromium.org/2726793002
Cr-Original-Commit-Position: refs/heads/master@{#454455}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54e4fbdde6ae7aa7844a0d11489ba76c9f859afb