ScriptStreamer: make state variables main-thread-only.
After this CL, |m_parsingFinished| is not set to true until the task posted
back to the main thread runs. After that, all access to the boolean variables
commented as protected by |ScriptStreamer::m_mutex| seems to be on the main
thread, so the mutex is also removed.

BUG=697651

Review-Url: https://codereview.chromium.org/2724153002
Cr-Original-Commit-Position: refs/heads/master@{#454568}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84609d29dbdaffc45415fc4b7ef6443096473d8b