Rearrange order of methods in baseline_optimizer.py and test.
Reason: In general, I think that it's easier to read classes when the
main, public methods are are above the helper, private methods.
Similarly, I think that in tests it's more readable if the test for main
public methods go above any tests of private helper methods.

This CL touches a lot of lines, but it only affects layout, order,
and comments, and should not affect behavior in any way.

Review-Url: https://codereview.chromium.org/2723253007
Cr-Original-Commit-Position: refs/heads/master@{#454731}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08e64091e1aff0f7efe873761fab7ace1a307fc5