Media Controls: make the MediaControlsOverlayPlayButton a runtime enabled flag.
It doesn't need to be a settings because it is not used on a document
basis but on a platform basis. A runtime enabled flag has a lower
overhead and is easier to work with because there is no need for a
settings object to be available, making tests easier to write.

BUG=698237

Review-Url: https://codereview.chromium.org/2728133002
Cr-Original-Commit-Position: refs/heads/master@{#454771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48a24fcc7bcf3f9c02faf6b1e949eed37f246c81