Media Controls: do not use events but C++ callbacks for insertion/removal from document.
For some reasons, the overhead of the events create some oilpan
allocated free memory. It's not a memory leak but it increases the
visible memory usage of Chrome for some perf bots. Using C++ callbacks
resolve the issue and should be fairly similar as far as the
functionality goes.

BUG=697980

Review-Url: https://codereview.chromium.org/2724983006
Cr-Original-Commit-Position: refs/heads/master@{#454772}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e679697d635baa92d4927b2b4e838f125266af1