Split transitions update storage between custom and standard properties
This patch creates a distinction between transition interpolations for
custom vs standard CSS properties.
This is required to support CSS Transitions on registered custom
properties as they need to be calculated and applied in separate phases.

This patch is a refactor and does not change behaviour.

BUG=671904

Review-Url: https://codereview.chromium.org/2724313002
Cr-Original-Commit-Position: refs/heads/master@{#454821}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 095ff780d828ee632f83ed54cfa9caf67179d7f0