Support reftest-wait in web-platform-tests
The web-platform-tests (wpt) suite allows tests to control when the screenshot is
taken for reftests, based on the presence (and then absence) of a 'reftest-wait'
class on the root element.

To support this mechanism a JavaScript shim is inserted into wpt tests, which uses
existing TestRunner methods (waitUntilDone, notifyDone), combined with a
MutationObserver on the root element.

BUG=698256

Review-Url: https://codereview.chromium.org/2729863004
Cr-Original-Commit-Position: refs/heads/master@{#455009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 255d25e6d389106b75c90110b4e00b022e472cc5