Fix orphaned wpt -expected.txt file deletion logic.
The previous logic was to use glob to check for corresponding
existing tests for a given baseline (-expected.txt) file, but
this is incorrect. The corresponding test for a given baseline
always involves adding one of a small set of possible extensions.

This CL also removes all files that would now be considered
orphaned baselines according to the new logic.

BUG=687746

Review-Url: https://codereview.chromium.org/2729133006
Cr-Original-Commit-Position: refs/heads/master@{#455017}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10ac6df8184863142d3fa9b60b0349b3d062978a