Re-land Catch up the spec update in CSS Rhythmic Sizing
This reverts r455020 and re-lands r455008 with a fix in core/BUILD.gn.

When files in core/css/properties are renamed, core/BUILD.gn needed to
be changed in addition to core/css/BUILD.gn, or the "compile confirm
no-op" step fails.

This can be confirmed locally by whether ninja prints
  ninja: no work to do.
line or not on the 2nd build in a row.

BUG=586413

Review-Url: https://codereview.chromium.org/2736043002
Cr-Original-Commit-Position: refs/heads/master@{#455337}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33bf66decbb5933a985bae31a2fa01e9fb46298f