Add initial BUILD.gn in platform/wtf/, and move first a few files there.
This is the first attempt to move files in wtf/ to platform/wtf/. In this
commit, initial BUILD.gn is introduced in platform/wtf/, and the following
files are moved to platform/wtf/ as a proof that the new BUILD.gn actually
works.

The following files are moved at this time:
    WTFExport.h
    CryptographicallyRandomNumber.{h,cpp}

They are chosen because they have very minimum dependency on other files
in wtf/, and we need at least one cpp file to make the target linkable.

BUG=691465

Review-Url: https://codereview.chromium.org/2703633003
Cr-Original-Commit-Position: refs/heads/master@{#455410}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00b8333a75c81b89aec8a377a33fa071a59e1748