Remove unused equality operator from ComputedStyleBase.
ComputedStyleBase::operator== is generated incorrectly since it includes
nonproperty fields as well. Fortunately, the equality operator is never
called anywhere, so it didn't cause any issues. This patch removes this
operator so that we don't accidentally use it in the future.

ComputedStyle has a correct equality implementation that should be used
instead.

BUG=628043

Review-Url: https://codereview.chromium.org/2736563002
Cr-Original-Commit-Position: refs/heads/master@{#455606}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 483a632594b00e8cbaa061aeb86c9f9fd965e2ba