Add common test configs for properties that take lengths.
This is a follow-up CL to https://codereview.chromium.org/2371673002, to make testing of length-accepting properties more complete.

To do this, I added "Length" as a typedom_types for each property in CSSProperties.in
While I was at it, I also added the valid keywords for those properties.

BUG=545318

Review-Url: https://codereview.chromium.org/2583063002
Cr-Original-Commit-Position: refs/heads/master@{#455687}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d77a388ceb8b2742c2ee4f0d2c8cc09a0aab4c6