Set "Running" state only when actually running
When resuming an AudioContext, the state should be set to "running"
when the context has actually started running again.  Previously, this
was set when resume() was called.  Depending on the device, it make
take a while before the device actually starts running, so we were
prematurely setting the state to "running".

This can't be tested in a layout test with offline context.

BUG=693978
TEST=none

Review-Url: https://codereview.chromium.org/2717613007
Cr-Original-Commit-Position: refs/heads/master@{#455867}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc4cda062ada67cd3c845149d58a4dc7545a00c1